+++ lib/keys/OpensslKeyPair.cpp
@@ -4,7 +4,7 @@
 #include <libcryptosec/DynamicEngine.h>
 #include "util/Util.h"
 
-JNIEXPORT jint JNICALL Java_br_ufsc_labsec_libcryptosec_keys_OpensslKeyPair__1init(JNIEnv* env, jobject obj, jint _engineReference, jstring _keyId)
+JNIEXPORT jlong JNICALL Java_br_ufsc_labsec_libcryptosec_keys_OpensslKeyPair__1init(JNIEnv* env, jobject obj, jlong _engineReference, jstring _keyId)
 {
 	std::string keyId(env->GetStringUTFChars(_keyId, 0));
 	DynamicEngine* engine = reinterpret_cast<DynamicEngine*>(_engineReference);
@@ -12,7 +12,7 @@ JNIEXPORT jint JNICALL Java_br_ufsc_labsec_libcryptosec_keys_OpensslKeyPair__1in
 	try
 	{
 		KeyPair* kp = new KeyPair(engine, keyId);
-		return reinterpret_cast<jint>(kp);
+		return reinterpret_cast<jlong>(kp);
 	}
 	catch(EngineException& ex)
 	{
@@ -22,11 +22,11 @@ JNIEXPORT jint JNICALL Java_br_ufsc_labsec_libcryptosec_keys_OpensslKeyPair__1in
 
 }
 
-JNIEXPORT jint JNICALL Java_br_ufsc_labsec_libcryptosec_keys_OpensslKeyPair__1getPublicKey(JNIEnv* env, jobject obj)
+JNIEXPORT jlong JNICALL Java_br_ufsc_labsec_libcryptosec_keys_OpensslKeyPair__1getPublicKey(JNIEnv* env, jobject obj)
 {
 	try
 	{
-		return reinterpret_cast<jint>(Util::getInstance<KeyPair*>(env, obj)->getPublicKey());
+		return reinterpret_cast<jlong>(Util::getInstance<KeyPair*>(env, obj)->getPublicKey());
 	}
 	catch(AsymmetricKeyException& ex)
 	{
@@ -40,11 +40,11 @@ JNIEXPORT jint JNICALL Java_br_ufsc_labsec_libcryptosec_keys_OpensslKeyPair__1ge
 	}
 }
 
-JNIEXPORT jint JNICALL Java_br_ufsc_labsec_libcryptosec_keys_OpensslKeyPair__1getPrivateKey(JNIEnv* env, jobject obj)
+JNIEXPORT jlong JNICALL Java_br_ufsc_labsec_libcryptosec_keys_OpensslKeyPair__1getPrivateKey(JNIEnv* env, jobject obj)
 {
 	try
 	{
-		return reinterpret_cast<jint>(Util::getInstance<KeyPair*>(env, obj)->getPrivateKey());
+		return reinterpret_cast<jlong>(Util::getInstance<KeyPair*>(env, obj)->getPrivateKey());
 	}
 	catch(AsymmetricKeyException& ex)
 	{
